+++ database/migrations/2023_03_10_115716_create_organizations_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('address');
             $table->integer('post_index');
-            $table->integer('organizers_id');
+            $table->unsignedBigInteger('organizers_id');
             $table->timestamps();
 
             $table->foreign('organizers_id')->references('organizers_id')
